Split up rich_text.py in rich_text and markdown
