+++ src/ultimate_notion/blocks.py
@@ -14,10 +14,11 @@ from typing_extensions import Self
 from ultimate_notion.comment import Comment, Discussion
 from ultimate_notion.core import InvalidAPIUsageError, NotionEntity, get_active_session, get_url
 from ultimate_notion.file import Emoji, FileInfo, wrap_icon
+from ultimate_notion.markdown import md_comment
 from ultimate_notion.obj_api import blocks as obj_blocks
 from ultimate_notion.obj_api import objects as objs
 from ultimate_notion.obj_api.enums import BGColor, CodeLang, Color
-from ultimate_notion.rich_text import Text, User, md_comment
+from ultimate_notion.rich_text import Text, User
 from ultimate_notion.utils import del_nested_attr
 
 if TYPE_CHECKING:
